+++ client/src/App.tsx
@@ -5,7 +5,7 @@ import { About } from './pages/About';
 import { CoursePage } from './pages/CoursePage';
 import { Explore } from './pages/Explore';
 import { Home } from './pages/Home';
-import { NotFound404 } from './pages/404';
+import { NotFound } from './pages/NotFound';
 
 const App = () => {
   return (
@@ -16,7 +16,7 @@ const App = () => {
       </Route>
       <Route path='/explore' element={<Explore />} />
       <Route path='/about' element={<About />} />
-      <Route path='/404' element={<NotFound404 />} />
+      <Route path='*' element={<NotFound />} />
     </Routes>
   );
 };
