+++ client/src/pages/NotFound.tsx
@@ -1,0 +1,13 @@
+import { Layout } from '../components/Layout';
+import { useNavigate } from 'react-router-dom';
+import { NotFound as NotFoundContent } from '../components/NotFound';
+
+export const NotFound = () => {
+  const nav = useNavigate();
+
+  return (
+    <Layout>
+      <NotFoundContent />
+    </Layout>
+  );
+};
